requirements.txt
Pillow
fastapi
uvicorn
rembg
